+++ packages/expo-notifications/plugin/src/__tests__/withNotificationsAndroid-test.ts
@@ -1,4 +1,4 @@
-import { ExpoConfig } from '@expo/config-types';
+import { ExpoConfig } from 'expo/config';
 import { fs, vol } from 'memfs';
 import * as path from 'path';