+++ packages/expo-splash-screen/CHANGELOG.md
@@ -14,6 +14,8 @@
 
 ### 💡 Others
 
+- [plugin] Migrate import from @expo/config-plugins to expo/config-plugins and @expo/config-types to expo/config. ([#18855](https://github.com/expo/expo/pull/18855) by [@brentvatne](https://github.com/brentvatne))
+
 ## 0.16.1 — 2022-07-16
 
 ### 🐛 Bug fixes